+++ webview-sys/build.rs
@@ -15,8 +15,7 @@ fn main() {
             .flag_if_supported("/std:c++17");
     } else {
         build
-            .include("webview_new.h")
-            .file("webview_gtk.c")
+            .include("webview.h")
             .flag_if_supported("-std=c11")
             .flag_if_supported("-w");
     }
@@ -29,6 +28,8 @@ fn main() {
 
     if target.contains("windows") {
         if !cfg!(feature = "edge") {
+            build.file("webview_mshtml.c");
+
             for &lib in &["ole32", "comctl32", "oleaut32", "uuid", "gdi32"] {
                 println!("cargo:rustc-link-lib={}", lib);
             }
@@ -39,13 +40,14 @@ fn main() {
             .probe("webkit2gtk-4.0")
             .unwrap();
 
+        build.file("webview_gtk.c");
+
         for path in webkit.include_paths {
             build.include(path);
         }
-        build.define("WEBVIEW_GTK", None);
     } else if target.contains("apple") {
         build
-            .define("WEBVIEW_COCOA", None)
+            .file("webview_cocoa.c")
             .define("OBJC_OLD_DISPATCH_PROTOTYPES", "1")
             .flag("-x")
             .flag("objective-c");
