+++ webview-sys/webview.h
@@ -1,26 +1,3 @@
-/*
- * MIT License
- *
- * Copyright (c) 2017 Serge Zaitsev
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
 #ifndef WEBVIEW_H
 #define WEBVIEW_H
 
@@ -35,89 +12,18 @@ extern "C" {
 #endif
 
 #include <stdint.h>
-#include <stdlib.h>
 #include <string.h>
-
-#if defined(WEBVIEW_GTK)
-
-#elif defined(WEBVIEW_WINAPI)
-
-#elif defined(WEBVIEW_COCOA)
-#include <objc/objc-runtime.h>
-#include <CoreGraphics/CoreGraphics.h>
-#include <limits.h>
-
-struct webview_priv {
-  id pool;
-  id window;
-  id webview;
-  id windowDelegate;
-  int should_exit;
-};
-#else
-#error "Define one of: WEBVIEW_GTK, WEBVIEW_COCOA or WEBVIEW_WINAPI"
-#endif
 
 struct webview;
 
-typedef void (*webview_external_invoke_cb_t)(struct webview *w,
-                                             const char *arg);
-
-struct webview {
-  const char *url;
-  const char *title;
-  int width;
-  int height;
-  int resizable;
-  int debug;
-  webview_external_invoke_cb_t external_invoke_cb;
-  struct webview_priv priv;
-  void *userdata;
-};
+typedef void (*webview_external_invoke_cb_t)(struct webview *w, const char *arg);
+typedef void (*webview_dispatch_fn)(struct webview *w, void *arg);
 
 enum webview_dialog_type {
   WEBVIEW_DIALOG_TYPE_OPEN = 0,
   WEBVIEW_DIALOG_TYPE_SAVE = 1,
   WEBVIEW_DIALOG_TYPE_ALERT = 2
 };
-
-#define WEBVIEW_DIALOG_FLAG_FILE (0 << 0)
-#define WEBVIEW_DIALOG_FLAG_DIRECTORY (1 << 0)
-
-#define WEBVIEW_DIALOG_FLAG_INFO (1 << 1)
-#define WEBVIEW_DIALOG_FLAG_WARNING (2 << 1)
-#define WEBVIEW_DIALOG_FLAG_ERROR (3 << 1)
-#define WEBVIEW_DIALOG_FLAG_ALERT_MASK (3 << 1)
-
-typedef void (*webview_dispatch_fn)(struct webview *w, void *arg);
-
-struct webview_dispatch_arg {
-  webview_dispatch_fn fn;
-  struct webview *w;
-  void *arg;
-};
-
-#define DEFAULT_URL                                                            \
-  "data:text/"                                                                 \
-  "html,%3C%21DOCTYPE%20html%3E%0A%3Chtml%20lang=%22en%22%3E%0A%3Chead%3E%"    \
-  "3Cmeta%20charset=%22utf-8%22%3E%3Cmeta%20http-equiv=%22X-UA-Compatible%22%" \
-  "20content=%22IE=edge%22%3E%3C%2Fhead%3E%0A%3Cbody%3E%3Cdiv%20id=%22app%22%" \
-  "3E%3C%2Fdiv%3E%3Cscript%20type=%22text%2Fjavascript%22%3E%3C%2Fscript%3E%"  \
-  "3C%2Fbody%3E%0A%3C%2Fhtml%3E"
-
-#define CSS_INJECT_FUNCTION                                                    \
-  "(function(e){var "                                                          \
-  "t=document.createElement('style'),d=document.head||document."               \
-  "getElementsByTagName('head')[0];t.setAttribute('type','text/"               \
-  "css'),t.styleSheet?t.styleSheet.cssText=e:t.appendChild(document."          \
-  "createTextNode(e)),d.appendChild(t)})"
-
-static const char *webview_check_url(const char *url) {
-  if (url == NULL || strlen(url) == 0) {
-    return DEFAULT_URL;
-  }
-  return url;
-}
 
 WEBVIEW_API int webview(const char *title, const char *url, int width,
                         int height, int resizable);
@@ -141,92 +47,48 @@ WEBVIEW_API void webview_exit(struct webview *w);
 WEBVIEW_API void webview_debug(const char *format, ...);
 WEBVIEW_API void webview_print_log(const char *s);
 
-#ifdef WEBVIEW_IMPLEMENTATION
-#undef WEBVIEW_IMPLEMENTATION
+WEBVIEW_API void* wrapper_webview_get_userdata(struct webview* w);
+WEBVIEW_API struct webview* wrapper_webview_new(const char* title, const char* url, int width, int height, int resizable, int debug, webview_external_invoke_cb_t external_invoke_cb, void* userdata);
+WEBVIEW_API void wrapper_webview_free(struct webview* w);
 
-WEBVIEW_API int webview(const char *title, const char *url, int width,
-                        int height, int resizable) {
-  struct webview webview;
-  memset(&webview, 0, sizeof(webview));
-  webview.title = title;
-  webview.url = url;
-  webview.width = width;
-  webview.height = height;
-  webview.resizable = resizable;
-  int r = webview_init(&webview);
-  if (r != 0) {
-    return r;
+#define WEBVIEW_DIALOG_FLAG_FILE (0 << 0)
+#define WEBVIEW_DIALOG_FLAG_DIRECTORY (1 << 0)
+
+#define WEBVIEW_DIALOG_FLAG_INFO (1 << 1)
+#define WEBVIEW_DIALOG_FLAG_WARNING (2 << 1)
+#define WEBVIEW_DIALOG_FLAG_ERROR (3 << 1)
+#define WEBVIEW_DIALOG_FLAG_ALERT_MASK (3 << 1)
+
+struct webview_dispatch_arg {
+  webview_dispatch_fn fn;
+  struct webview *w;
+  void *arg;
+};
+
+#define DEFAULT_URL                                                            \
+  "data:text/"                                                                 \
+  "html,%3C%21DOCTYPE%20html%3E%0A%3Chtml%20lang=%22en%22%3E%0A%3Chead%3E%"    \
+  "3Cmeta%20charset=%22utf-8%22%3E%3Cmeta%20http-equiv=%22X-UA-Compatible%22%" \
+  "20content=%22IE=edge%22%3E%3C%2Fhead%3E%0A%3Cbody%3E%3Cdiv%20id=%22app%22%" \
+  "3E%3C%2Fdiv%3E%3Cscript%20type=%22text%2Fjavascript%22%3E%3C%2Fscript%3E%"  \
+  "3C%2Fbody%3E%0A%3C%2Fhtml%3E"
+
+#define CSS_INJECT_FUNCTION                                                    \
+  "(function(e){var "                                                          \
+  "t=document.createElement('style'),d=document.head||document."               \
+  "getElementsByTagName('head')[0];t.setAttribute('type','text/"               \
+  "css'),t.styleSheet?t.styleSheet.cssText=e:t.appendChild(document."          \
+  "createTextNode(e)),d.appendChild(t)})"
+
+static inline const char *webview_check_url(const char *url) {
+  if (url == NULL || strlen(url) == 0) {
+    return DEFAULT_URL;
   }
-  while (webview_loop(&webview, 1) == 0) {
-  }
-  webview_exit(&webview);
-  return 0;
+  return url;
 }
-
-WEBVIEW_API void webview_debug(const char *format, ...) {
-  char buf[4096];
-  va_list ap;
-  va_start(ap, format);
-  vsnprintf(buf, sizeof(buf), format, ap);
-  webview_print_log(buf);
-  va_end(ap);
-}
-
-static int webview_js_encode(const char *s, char *esc, size_t n) {
-  int r = 1; /* At least one byte for trailing zero */
-  for (; *s; s++) {
-    const unsigned char c = *s;
-    if (c >= 0x20 && c < 0x80 && strchr("<>\\'\"", c) == NULL) {
-      if (n > 0) {
-        *esc++ = c;
-        n--;
-      }
-      r++;
-    } else {
-      if (n > 0) {
-        snprintf(esc, n, "\\x%02x", (int)c);
-        esc += 4;
-        n -= 4;
-      }
-      r += 4;
-    }
-  }
-  return r;
-}
-
-WEBVIEW_API int webview_inject_css(struct webview *w, const char *css) {
-  int n = webview_js_encode(css, NULL, 0);
-  char *esc = (char *)calloc(1, sizeof(CSS_INJECT_FUNCTION) + n + 4);
-  if (esc == NULL) {
-    return -1;
-  }
-  char *js = (char *)calloc(1, n);
-  webview_js_encode(css, js, n);
-  snprintf(esc, sizeof(CSS_INJECT_FUNCTION) + n + 4, "%s(\"%s\")",
-           CSS_INJECT_FUNCTION, js);
-  int r = webview_eval(w, esc);
-  free(js);
-  free(esc);
-  return r;
-}
-
-#if defined(WEBVIEW_GTK)
-
-#endif /* WEBVIEW_GTK */
-
-#if defined(WEBVIEW_WINAPI)
-
-#endif /* WEBVIEW_WINAPI */
-
-#if defined(WEBVIEW_COCOA)
-
-
-#endif /* WEBVIEW_COCOA */
-
-#endif /* WEBVIEW_IMPLEMENTATION */
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* WEBVIEW_H */
+#endif // WEBVIEW_H
