+++ webview-sys/webview_cocoa.c
@@ -1,4 +1,4 @@
-#include "webview_new.h"
+#include "webview.h"
 
 #include <objc/objc-runtime.h>
 #include <CoreGraphics/CoreGraphics.h>
