+++ webview-sys/webview_mshtml.c
@@ -1,4 +1,4 @@
-#include "webview_new.h"
+#include "webview.h"
 
 #define CINTERFACE
 #include <windows.h>
